Port のファイル位置を変更する api を追加
+ scm_capi_port_seek を追加
+ scm_port_seek のバグを修正

+++ src/include/scythe/port.h
@@ -271,7 +271,7 @@ ssize_t scm_port_peek_bytes(ScmObj port, void *buf, size_t size);
 ssize_t scm_port_peek_char(ScmObj port, scm_char_t *chr);
 ssize_t scm_port_write_bytes(ScmObj port, const void *buf, size_t size);
 ssize_t scm_port_write_char(ScmObj port, scm_char_t chr);
-int scm_port_seek(ScmObj port, off_t offset, int whence);
+off_t scm_port_seek(ScmObj port, off_t offset, int whence);
 off_t scm_port_pos(ScmObj port);
 const void *scm_port_string_buffer(ScmObj port);
 ssize_t scm_port_string_buffer_length(ScmObj port);
